pyproject.toml
[project]
name = "cis-lab-smart-lock"
version = "0.1.0"
description = "学生証から学生番号を読み取って研究室の鍵を開け閉めするプログラム"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nfcpy>=1.0.4",
    "python-dotenv>=1.0.1",
    "pyusb>=1.3.1",
    "requests>=2.32.3",
    "slack-sdk>=3.34.0",
]

[tool.isort]
profile = "google"
